Starting to work on photometric calibration.  In so doing discovered a problem with source lists: SEXtractor apertures are specified in *diameters*, but we use radii, and we've been assuming that SExtractor uses radii.  Fix that.  Also, fix PSF-exists handling in detection.py so that it really gets SExtractor PSF photometry (whatever those numbers actually mean).  Remove the "ra" and "dec" property from source list, and remove the *_WORLD parmeters from what SExtractor extracts, because source lists are created before we have a good WCS.

+++ pipeline/detection.py
@@ -246,12 +246,12 @@ class Detector:
         tempnamebase = ''.join( random.choices( 'abcdefghijklmnopqrstuvwxyz', k=10 ) )
         sourcepath = pathlib.Path( FileOnDiskMixin.temp_path ) / f'{tempnamebase}.sources.fits'
         psfpath = pathlib.Path( psffile ) if psffile is not None else None
-        psfxmlpath = pathlib.Path( FileOnDiskMixin.temp_path ) / f'{tempnamebase}.sources.psf.xml'
+        psfxmlpath = None
 
         apers = ( np.array( self.pars.apers ) if self.pars.apers is not None
                   else np.array( [1., 2., 3., 4., 5., 7., 10.] ) )
 
-        if self.pars.measure_psf or ( self.pars.psf is None ):
+        if self.pars.measure_psf:
             # Run sextractor once without a psf to get objects from
             # which to build the psf.
             #
@@ -267,14 +267,14 @@ class Detector:
 
             # Get the PSF
             psf = self._run_psfex( tempnamebase, image.id, do_not_cleanup=True )
+            psfpath = pathlib.Path( FileOnDiskMixin.temp_path ) / f'{tempnamebase}.sources.psf'
+            psfxmlpath = pathlib.Path( FileOnDiskMixin.temp_path ) / f'{tempnamebase}.sources.psf.xml'
             psf.image = image
-        else:
+        elif self.pars.psf is not None:
             psf = self.pars.psf
-            if psf is not None:
-                psfpath, psfxmlpath = self.psf.get_fullpath( download=download, always_verify_md5=True )
-            else:
-                psfpath = None
-                psfxmlpath = None
+            psfpath, psfxmlpath = psf.get_fullpath()
+        else:
+            psf = None
 
         if self.pars.aperunit == 'fwhm':
             if psf is None:
@@ -296,7 +296,7 @@ class Detector:
 
         # Clean up the temporary files created (that weren't already cleaned up by _run_sextractor_once)
         sourcepath.unlink( missing_ok=True )
-        if self.pars.psf is None:
+        if ( psffile is None ) and ( self.pars.psf is None ):
             if psfpath is not None: psfpath.unlink( missing_ok=True )
             if psfxmlpath is not None: psfxmlpath.unlink( missing_ok=True )
 
@@ -470,7 +470,7 @@ class Detector:
                      "-WEIGHT_GAIN", "N",
                      "-FLAG_IMAGE", str(tmpflags),
                      "-FLAG_TYPE", "OR",
-                     "-PHOT_APERTURES", ",".join( [ str(a) for a in apers ] ),
+                     "-PHOT_APERTURES", ",".join( [ str(a*2.) for a in apers ] ),
                      "-SATUR_LEVEL", str( image.instrument_object.average_saturation_limit( image ) ),
                      "-STARNNW_NAME", nnw,
                      "-BACK_TYPE", "AUTO",
